[UPDATE]: Methods Updates for Category and by Date

+++ src/main/java/com/finance/DTO/ExpenseDTO.java
@@ -6,16 +6,19 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
+import java.time.LocalDateTime;
+
 @Data
 @NoArgsConstructor
 public class ExpenseDTO {
-    private Long amount;
+    private BigDecimal amount;
     private String category;
     private String description;
     private LocalDate date;
 
-    public ExpenseDTO(Long amount, String category, String description, LocalDate date) {
+    public ExpenseDTO(BigDecimal amount, String category, String description, LocalDate date) {
     this.amount=amount;
     this.date=date;
     this.category=category;
